+++ src/components/Publish/index.js
@@ -1,7 +1,6 @@
 
 import React, { useContext } from "react";
 import { func } from "prop-types";
-import { CopyToClipboard } from "react-copy-to-clipboard";
 import ModalContainer from "common/ModalContainer";
 import NextButton from "common/NextButton";
 import MnemonicContext from "context/MnemonicContext";
@@ -10,16 +9,14 @@ import usericon from 'assets/users.svg'
 import emailicon from 'assets/email.svg'
 import birthicon from 'assets/birthicon.svg'
 import twittericon from 'assets/twittericon.svg'
-import { toast, Slide } from "react-toastify";
 import "react-toastify/dist/ReactToastify.css";
 import GetDids from "services/getdids.service"
 import './index.scss';
 
-var QRCode = require("qrcode.react");
 
 
 function Publish({ setStep }) {
-  const { did, name, email, birthDate, twitter_user, setPublishStatus, mnemonic } = useContext(MnemonicContext)
+  const { name, email, birthDate, twitter_user, setPublishStatus, mnemonic } = useContext(MnemonicContext)
 
   const publishDocument = async () => {
     let confirmation = await GetDids.PublishDocument(mnemonic, {
@@ -29,24 +26,11 @@ function Publish({ setStep }) {
       twitter: twitter_user
     })
 
-    //let confirmation = "asdasdd"
     
     setPublishStatus(confirmation)
     setStep()
   }
-
-  const notify = () =>
-    toast.info("☑ DID successfully copied to clipboard", {
-      className: "bg-blue-600",
-      position: "top-right",
-      autoClose: 3000,
-      hideProgressBar: true,
-      closeOnClick: true,
-      pauseOnHover: false,
-      draggable: false,
-      transition: Slide,
-      progress: undefined,
-    });
+  
   return (
     <ModalContainer>
       <span className="title">Nearly completed!</span>
